- Fix `QuadraticEquation`'s and `CubicEquation`'s `calculate()` method. - Fix `factorizate()` method of `CompositeNumber` class. - Implement `*`, `/`, `+` and `-` operators for complex number `I`. - Add `module`, `argument` getters to `I` class. - Changed `double` type of elements to `num` type. - Override `==` operator for `MatrixBase`, `VectorBase` and `Complex` classes.

+++ lib/src/discrete_mathematics/number_theory/primes.dart
@@ -1,5 +1,5 @@
 /// Prime numbers from 2 to 271
-final List<double> primes = <double>[
+final List<int> primes = <int>[
   2,
   3,
   5,
